+++ src/Tests/Measuremap.IntegrationTest/OnStartTests.cs
@@ -2,15 +2,15 @@
 
 namespace Measuremap.IntegrationTest
 {
-    public class ExecutionContextTests
+    public class OnStartTests
     {
         [Test]
-        public void ExecutionContext_MultyThreadSessionHandler()
+        public void OnStart_MultyThreadSessionHandler()
         {
             var context = new ExecutionContext();
 
             ProfilerSession.StartSession()
-                .CreateExecutionContext(s => context)
+                .OnStart(s => context)
                 .Task(c => c.Should().BeSameAs(context))
                 .SetThreads(1)
                 .RunWarmup(false)
@@ -18,25 +18,25 @@ namespace Measuremap.IntegrationTest
         }
 
         [Test]
-        public void ExecutionContext_Warmup()
+        public void OnStart_Warmup()
         {
             var context = new ExecutionContext();
 
             ProfilerSession.StartSession()
-                .CreateExecutionContext(s => context)
+                .OnStart(s => context)
                 .Task(c => c.Should().BeSameAs(context))
                 .SetThreads(1)
                 .RunSession();
         }
 
         [Test]
-        public void ExecutionContext_MainThreadSessionHandler()
+        public void OnStart_MainThreadSessionHandler()
         {
             var context = new ExecutionContext();
             context.Set("test", 1);
 
             ProfilerSession.StartSession()
-                .CreateExecutionContext(s => context)
+                .OnStart(s => context)
                 .Task(c => c.Get<int>("test").Should().Be(1))
                 .SetThreadBehaviour(ThreadBehaviour.RunOnMainThread)
                 .RunWarmup(false)
@@ -45,12 +45,12 @@ namespace Measuremap.IntegrationTest
         }
 
         [Test]
-        public void ExecutionContext_Basic()
+        public void OnStart_Basic()
         {
             var context = new ExecutionContext();
 
             ProfilerSession.StartSession()
-                .CreateExecutionContext(s => context)
+                .OnStart(s => context)
                 .Task(c => c.Should().BeSameAs(context))
                 .RunWarmup(false)
                 .SetIterations(1)